KAFKA-20553 Move RocksDBStoreTestingUtils to streams:integration-tests
